--- client/src/App.js
@@ -1,48 +1,0 @@
-import React, { useState, useEffect, useRef } from 'react';
-import './App.css';
-
-function App() {
-  const [question, setQuestion] = useState('');
-  const [conversation, setConversation] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const chatEndRef = useRef(null);
-
-  const backendUrl = process.env.REACT_APP_BACKEND_URL || 'http://localhost:5000';
-
-  // Function to submit question to the server
-  const askQuestion = async () => {
-    // TODO: Implement the function to send the question to the server
-  };
-
-  return (
-    <div className="container">
-      <h1>Chat with me</h1>
-
-      {/* Chat container */}
-      <div className="chat-container">
-        {conversation.map((msg, index) => (
-          <div key={index} className={`message ${msg.role}`}>
-            <strong>{msg.role === 'user' ? 'You' : 'GPT-4'}:</strong>
-            <span>{msg.content}</span>
-          </div>
-        ))}
-        <div ref={chatEndRef} />
-      </div>
-
-      {/* Input section */}
-      <textarea
-        className="textarea"
-        rows="4"
-        placeholder="Ask a question..."
-        value={question}
-        onChange={(e) => setQuestion(e.target.value)}
-      />
-      <br />
-      <button className="button" onClick={askQuestion} disabled={loading}>
-        {loading ? 'Loading...' : 'Ask'}
-      </button>
-    </div>
-  );
-}
-
-export default App;
