+++ client/src/Error/handleErrors.jsx
@@ -1,0 +1,7 @@
+const handleError = (err) => {
+  console.log("Handling error:", err);
+
+  // document.render(<div>Hi</div>);
+};
+
+export default handleError;
