+++ client/src/pages/Logout.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../context/LoginState";
+import { useAuth } from "../context/AuthContext";
 
 function Logout() {
   const navigate = useNavigate();
