+++ server/src/controllers/chat.controller.js
@@ -14,12 +14,20 @@ const handleLoggedInUser = async (
 ) => {
   console.log("authToken recieved");
 
+  if (!sessionToken) {
+    console.log("SessionToken not provided to store chat");
+    return res
+      .json(400)
+      .json({ msg: "sessionToken not provided to store chat" });
+  }
+
   // Find user with authToken
   jwt.verify(
     authToken,
     process.env.JWT_TOKEN_SECRET,
     async function (err, decoded) {
       if (err) {
+        console.log("Error occured while finding user with authToken");
         return res.status(500).json({ msg: "Error fetching in user" });
       }
 
@@ -27,6 +35,7 @@ const handleLoggedInUser = async (
 
       // If no user found
       if (!user) {
+        console.log("No user found");
         return res
           .status(400)
           .json({ msg: "You are not logged in to store chat" });
@@ -69,6 +78,8 @@ const handleChat = async (req, res) => {
   const { prompt, sessionToken } = req.body;
   const authToken = req.cookies.authToken;
 
+  console.log("Prompt:", prompt);
+
   if (!prompt) {
     return res.status(400).json({ msg: "Provide prompt" });
   }
@@ -78,12 +89,11 @@ const handleChat = async (req, res) => {
   ).getGenerativeModel({ model: "gemini-1.5-flash" });
 
   const result = await model.generateContent(prompt);
-  // const response = result.response.text();
 
   // If user not logged in
   if (!authToken) {
-    console.log("No sessionToken provided");
-    res.status(200).json({ response: result.response.text() });
+    console.log("authToken not provided");
+    return res.status(200).json({ response: result.response.text() });
   } else {
     await handleLoggedInUser(res, prompt, sessionToken, authToken, result);
   }
@@ -95,6 +105,7 @@ const handleChatWithFile = async (req, res) => {
   const file = req.file;
 
   const fileManager = new GoogleAIFileManager(process.env.GEMINI_AI_API_KEY);
+
   // Upload file to API Server
   const uploadResult = await fileManager.uploadFile(`${file.path}`, {
     mimeType: file.mimetype,
@@ -109,6 +120,7 @@ const handleChatWithFile = async (req, res) => {
   fs.unlink(file.path, (err) => {
     if (err) {
       console.log("Error deleting file:", err);
+      return res.status(500).json({ msg: "Server error in deleting file" });
     } else {
       console.log("File deleted successfully");
     }
@@ -131,7 +143,7 @@ const handleChatWithFile = async (req, res) => {
 
   // If user not logged in
   if (!authToken) {
-    console.log("No sessionToken provided");
+    console.log("authToken not provided");
     res.status(200).json({ response: result.response.text() });
   } else {
     await handleLoggedInUser(res, prompt, sessionToken, authToken, result);
